Commit changes. (See readme)
Changes mentioned in the readme are: This version makes the guess process non-case-sensitive, removes the need for the conio.h library, which is non-standard and outdated, tells the user what letters they have guessed wrong and is more stable because putting in more than 20 characters won't crash the program anymore. (The limit has not been removed, I think the original author wanted it to be simple and fast. The program won't crash, but it won't let you play properly either.)